+++ ruststep/src/parser/exchange/record.rs
@@ -1,4 +1,4 @@
-use crate::parser::exchange::*;
+use crate::step::*;
 use serde::{de, forward_to_deserialize_any};
 
 /// A struct typed in EXPRESS schema
