+++ ruststep/src/step/mod.rs
@@ -1,7 +1,9 @@
 //! Primitive data structures in STEP file
 
 mod logical;
+mod parameter;
 mod value;
 
 pub use logical::*;
+pub use parameter::*;
 pub use value::*;
