+++ ruststep/src/tables.rs
@@ -49,7 +49,7 @@
 //!     - `a_iter()` as above example
 //!
 
-use crate::{parser::exchange::Parameter, step::*};
+use crate::step::*;
 use serde::{
     de::{self, IntoDeserializer, VariantAccess},
     Deserialize,
